+++ packages/orq-rc/src/lib/config.ts
@@ -58,7 +58,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "2.0",
-  sdkVersion: "4.11.0-rc.17",
-  genVersion: "2.903.5",
-  userAgent: "speakeasy-sdk/typescript 4.11.0-rc.17 2.903.5 2.0 @orq-ai/node",
+  sdkVersion: "4.11.0-rc.18",
+  genVersion: "2.904.2",
+  userAgent: "speakeasy-sdk/typescript 4.11.0-rc.18 2.904.2 2.0 @orq-ai/node",
 } as const;