+++ packages/orq-rc/src/models/operations/invokeeval.ts
@@ -215,6 +215,7 @@ export type ResponseBodyJury = {
 export type InvokeEvalResponseBodyEvalsValue = {
   workflowRunId?: string | undefined;
   traceId?: string | undefined;
+  spanId?: string | undefined;
   value: number | boolean | string | null;
   explanation?: string | null | undefined;
   originalValue?: number | boolean | string | null | undefined;
@@ -858,6 +859,7 @@ export const InvokeEvalResponseBodyEvalsValue$inboundSchema: z.ZodType<
 > = z.object({
   workflow_run_id: z.string().optional(),
   trace_id: z.string().optional(),
+  span_id: z.string().optional(),
   value: z.nullable(z.union([z.number(), z.boolean(), z.string()])),
   explanation: z.nullable(z.string()).optional(),
   original_value: z.nullable(z.union([z.number(), z.boolean(), z.string()]))
@@ -868,6 +870,7 @@ export const InvokeEvalResponseBodyEvalsValue$inboundSchema: z.ZodType<
   return remap$(v, {
     "workflow_run_id": "workflowRunId",
     "trace_id": "traceId",
+    "span_id": "spanId",
     "original_value": "originalValue",
     "original_explanation": "originalExplanation",
   });