+++ packages/orq-rc/src/models/operations/updatedataset.ts
@@ -182,7 +182,7 @@ export const UpdateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-06-12T11:25:38.749Z",
+    "2026-06-14T13:29:29.915Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {