+++ packages/orq-rc/RELEASES.md
@@ -9959,3 +9959,13 @@ Based on:
 - [typescript v4.11.0-rc.17] packages/orq-rc
 ### Releases
 - [NPM v4.11.0-rc.17] https://www.npmjs.com/package/@orq-ai/node/v/4.11.0-rc.17 - packages/orq-rc
+
+## 2026-06-14 13:32:13
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.778.0 (2.904.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.11.0-rc.18] packages/orq-rc
+### Releases
+- [NPM v4.11.0-rc.18] https://www.npmjs.com/package/@orq-ai/node/v/4.11.0-rc.18 - packages/orq-rc